Add 2024 day 13 part 2

+++ 2024/Day13/Day13/Entities.cs
@@ -1,7 +1,7 @@
 namespace Day13;
 
-public record ButtonOffset(int Dx, int Dy);
+public record ButtonOffset(long Dx, long Dy);
 public record Button(char Id, ButtonOffset Offset, int TokenCost);
-public record Location(int X, int Y);
+public record Location(long X, long Y);
 
 public record MachineBehaviour(List<Button> Buttons, Location PrizeLocation);
